+++ jest.config.cjs
@@ -1,0 +1,6 @@
+module.exports = {
+  testMatch: ['**/?(*.)+(test).+(ts)'],
+  transform: {
+    '^.+\\.(ts|tsx)$': 'ts-jest'
+  }
+};