+++ src/index.ts
@@ -1,0 +1,3 @@
+/*! multi-integer-range (c) 2015-2021 Soichiro Miki */
+export * from './fp';
+export * from './MultiRange';